+++ src/MainApp.h
@@ -9,13 +9,29 @@
 
 #include "res\resource.h"
 
-class CMainApp : public CWinApp
+class CMainApp : public CWinAppEx
 {
 public:
     CMainApp();
 public:
     virtual BOOL InitInstance();
     DECLARE_MESSAGE_MAP()
+public:
+    bool m_bIsPortable = true;
+public:
+    CString CombinePath(CString szPath, CString szFile);
+public:
+    CString szSettingsPath;
+public:
+    CString szFormatsPath;
+    CString szLanguagesPath;
+    CString szProgressPath;
+    CString szToolsPath;
+public:
+    CString szOptionsFile;
+    CString szFormatsFile;
+    CString szItemsFile;
+    CString szToolsFile;
 };
 
 extern CMainApp mainApp;