+++ src/StdAfx.h
@@ -25,10 +25,16 @@
 
 #include <afxwin.h>
 #include <afxext.h>
+#include <afxdisp.h>
+
+#ifndef _AFX_NO_OLE_SUPPORT
 #include <afxdtctl.h>
+#endif
 #ifndef _AFX_NO_AFXCMN_SUPPORT
 #include <afxcmn.h>
 #endif
+
+#include <afxcontrolbars.h> 
 
 #ifdef _UNICODE
 #if defined _M_IX86